Vendor: Add kernel upstream script for maintainers
This is a script designed to help maintainers keep their kernels
easily up to date! Some major features:

* Gets the latest upstream version from kernel.org automatically
* Can cherry-pick or merge based on your preference
* Will error out if it detects conflicts you need to fix
* Gives you three update options
  - Updates to the next direct version (default)
  - Updates to a specified version
  - Updates to the latest version

For more information about the script, run "./kernel-upstream.sh -h"

Alex's note -

Quick example of how to use this for those who just want to merge in the
latest

- Make sure you type in ". build/envsetup.sh" and all that jazz

- Just cd into vendor/du/scripts

- Type in ./kernel-upstream.sh -mg -sf /OEM/KERNEL_NAME

you may have to place bash in front of the command above depending on your
build server or machine.

So for me I just had to type in bash ./kernel-upstream.sh -mg -sf /oneplus/msm8996
and it brought in the latest.

Change-Id: I0fefc67bf893075cdfc2630c5e9336658267ff4d