Compile WallpaperPicker for everyone
Currently this is only set to compile by inheriting the sdk base
and on a handful of devices (mainly Nexus and Pixel).

This change also resolves two issues. The first one being the issue
of users wanting to revert back to the default wallpaper and the second
one being unable to set wallpapers to the lockscreen/homescreen only
without using 3rd party apps.

Change-Id: I4fee0d543bd1f5c2f00e5bc91b48576123a93a39